+++ gensio/gensio_filter_telnet.c
@@ -172,8 +172,8 @@ struct telnet_buffer_data {
 };
 
 static int
-telnet_buffer_do_write(void *cb_data, void *buf, size_t buflen,
-		       size_t *written)
+telnet_buffer_do_write(void *cb_data, void *buf, unsigned int buflen,
+		       unsigned int *written)
 {
     struct telnet_buffer_data *data = cb_data;
     unsigned int count;
